discord_bot/pyproject.toml
[project]
name = "minecraft-discord-bot"
version = "1.0.0"
description = "Discord bot for Minecraft chat sync and server status"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.3.0",
    "mcrcon>=0.7.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["bot", "config"]
omit = ["tests/*"]
